--- preprocesado/conexionBBDD/constantesBBDD.py
@@ -1,3 +1,0 @@
-BBDD_URL = 'BBDD_URL'
-YT_CHANNELS = "youtube_data//youtube_files//yt_channels_info.csv"
-YT_VIDEOS = "youtube_data//youtube_files//yt_videos_info.csv"
